requirements.txt
aniso8601==1.3.0
beautifulsoup4==4.6.0
bleach==2.1.1
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
colorama==0.3.9
cycler==0.10.0
decorator==4.1.2
dj-database-url==0.4.2
Django==1.11.7
djangorestframework==3.7.1
entrypoints==0.2.3
flake8==3.5.0
Flask==0.12.2
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
graphviz==0.8.1
gunicorn==19.7.1
html5lib==1.0b10
idna==2.6
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.5
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
lxml==4.0.0
MarkupSafe==1.0
matplotlib==2.1.0
mccabe==0.6.1
mistune==0.8.1
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.1
numpy==1.13.3
olefile==0.44
pandas==0.21.0
pandas-datareader==0.5.0
pandocfilters==1.4.2
parso==0.1.0
pew==1.1.0
pickleshare==0.7.4
Pillow==4.3.0
pipenv==8.3.2
prompt-toolkit==1.0.15
psutil==5.3.1
psycopg2==2.7.3.2
pycodestyle==2.3.1
pydot==1.2.3
pydotplus==2.0.2
pyflakes==1.6.0
Pygments==2.2.0
pymongo==3.5.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.18.4
requests-file==1.4.2
requests-ftp==0.3.1
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.1.14
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
whitenoise==3.3.1
widgetsnbextension==3.0.8
